requirements.txt
-i https://pypi.org/simple
cycler==0.10.0
ffmpeg-python==0.2.0
future==0.18.2
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
prompt-toolkit==3.0.5
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.4.1
six==1.15.0
wcwidth==0.2.2
youtube-dl==2020.5.29
